Changed function creatingInputOnText to reusable
